refactor: delete column commad
