+++ packages/editor/src/core/extensions/table/table/table.ts
@@ -7,7 +7,6 @@ import {
   addRowBefore,
   CellSelection,
   columnResizing,
-  deleteColumn,
   deleteRow,
   deleteTable,
   fixTables,
@@ -27,6 +26,7 @@ import { TableInsertPlugin } from "../plugins/insert-handlers/plugin";
 import { tableControls } from "./table-controls";
 import { TableView } from "./table-view";
 import { createTable } from "./utilities/create-table";
+import { deleteColumnOrTable } from "./utilities/delete-column";
 import { deleteTableWhenAllCellsSelected } from "./utilities/delete-table-when-all-cells-selected";
 import { insertLineAboveTableAction } from "./utilities/insert-line-above-table-action";
 import { insertLineBelowTableAction } from "./utilities/insert-line-below-table-action";
@@ -140,40 +140,7 @@ export const Table = Node.create<TableOptions>({
         () =>
         ({ state, dispatch }) =>
           addColumnAfter(state, dispatch),
-      deleteColumn:
-        () =>
-        ({ state, dispatch }) => {
-          const { selection } = state;
-
-          // Check if we're in a table and have a cell selection
-          if (!(selection instanceof CellSelection)) {
-            return false;
-          }
-
-          // Get the ProseMirrorTable and calculate total columns
-          const tableStart = selection.$anchorCell.start(-1);
-          const selectedTable = state.doc.nodeAt(tableStart - 1);
-
-          if (!selectedTable) return false;
-
-          // Count total columns by examining the first row
-          const firstRow = selectedTable.firstChild;
-          if (!firstRow) return false;
-
-          let totalColumns = 0;
-          for (let i = 0; i < firstRow.childCount; i++) {
-            const cell = firstRow.child(i);
-            totalColumns += cell.attrs.colspan || 1;
-          }
-
-          // If only one column exists, delete the entire ProseMirrorTable
-          if (totalColumns === 1) {
-            return deleteTable(state, dispatch);
-          }
-
-          // Otherwise, proceed with normal column deletion
-          return deleteColumn(state, dispatch);
-        },
+      deleteColumn: deleteColumnOrTable,
       addRowBefore:
         () =>
         ({ state, dispatch }) =>
